+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog].
   default is t) ([#261]).
 
 ### Enhancements
+* The prompt gets selected when using `next-history-element` and the
+  prompt equals the default ([#323], [#324]).
 * Computation of candidates is faster for `describe-variable` ([#312],
   [#316], [#320], [#321]).
 * Candidates of `completing-read-multiple` which are submitted by
@@ -69,6 +71,9 @@ The format is based on [Keep a Changelog].
 * When let binding `minibuffer-message-timeout` around
   `minibuffer-message` the value wouldn't be applied, which has been
   fixed ([#327]).
+* `minibuffer-default` is now treated as the default when set, before
+  it would have no effect. When a list the car is used as default as
+  of now ([#324]).
 * `selectrum-extend-current-candidate-highlight`,
   `selectrum-show-indices`, `selectrum-right-margin-padding` and
   `selectrum-multiline-display-settings` wouldn't use the local
@@ -158,6 +163,8 @@ The format is based on [Keep a Changelog].
 [#318]: https://github.com/raxod502/selectrum/pull/318
 [#320]: https://github.com/raxod502/selectrum/issues/320
 [#321]: https://github.com/raxod502/selectrum/pull/321
+[#323]: https://github.com/raxod502/selectrum/issues/323
+[#324]: https://github.com/raxod502/selectrum/pull/324
 [#327]: https://github.com/raxod502/selectrum/pull/327
 
 ## 3.0 (released 2020-10-20)